refactor: centralize combat logic
